+++ packages/cli/src/utilities/format.ts
@@ -1,12 +1,24 @@
 import prettier from 'prettier';
+import {extname} from 'path';
 
-const PRETTIER_CONFIG = {
-  ...require('@shopify/prettier-config'),
-};
+const DEFAULT_PRETTIER_CONFIG = {...require('@shopify/prettier-config')};
 
-export function formatFile(content: string) {
-  // TODO: Search for local project config with fallback to Shopify
-  const formattedContent = prettier.format(content, PRETTIER_CONFIG);
+export async function formatFile(path: string, content: string) {
+  const ext = extname(path);
+  const prettierConfig = {
+    // TODO: Search for local project config with fallback to Shopify
+    ...DEFAULT_PRETTIER_CONFIG,
+    parser: 'babel',
+  };
+
+  switch (ext) {
+    case '.html':
+    case '.css':
+      prettierConfig.parser = ext.slice(1);
+      break;
+  }
+
+  const formattedContent = await prettier.format(content, prettierConfig);
 
   return formattedContent;
 }