+++ packages/hydrogen/src/hooks/useShopQuery/hooks.ts
@@ -19,6 +19,7 @@ export function useShopQuery<T>({
   query,
   variables = {},
   cache = {},
+  locale = '',
 }: {
   /** A string of the GraphQL query.
    * If no query is provided, useShopQuery will make no calls to the Storefront API.
@@ -28,6 +29,8 @@ export function useShopQuery<T>({
   variables?: Record<string, any>;
   /** An object containing cache-control options for the sub-request. */
   cache?: CacheOptions;
+  /** A string corresponding to a valid locale identifier like `en-us` used to make the request. */
+  locale?: string;
 }): UseShopQueryResponse<T> {
   if (isClient()) {
     throw new Error(
@@ -36,7 +39,7 @@ export function useShopQuery<T>({
   }
 
   const body = query ? graphqlRequestBody(query, variables) : '';
-  const {request, key} = createShopRequest(body);
+  const {request, key} = createShopRequest(body, locale);
 
   const {data, error: fetchError} = useQuery<UseShopQueryResponse<T>>(
     key,
@@ -89,8 +92,13 @@ export function useShopQuery<T>({
   return data as UseShopQueryResponse<T>;
 }
 
-function createShopRequest(body: string) {
-  const {storeDomain, storefrontToken, graphqlApiVersion} = useShop();
+function createShopRequest(body: string, locale?: string) {
+  const {
+    storeDomain,
+    storefrontToken,
+    graphqlApiVersion,
+    locale: defaultLocale,
+  } = useShop();
 
   const url = `https://${storeDomain}/api/${graphqlApiVersion}/graphql.json`;
 
@@ -100,9 +108,10 @@ function createShopRequest(body: string) {
       headers: {
         'X-Shopify-Storefront-Access-Token': storefrontToken,
         'content-type': 'application/json',
+        'Accept-Language': (locale as string) ?? defaultLocale,
       },
       body,
     }),
-    key: [storeDomain, graphqlApiVersion, body],
+    key: [storeDomain, graphqlApiVersion, body, locale],
   };
 }